Fix edge case with ejs subdir
Ejs doesn't have a package directory, it instead has a named folder. In
order to deal with this, read the contents of where the tarball was
unpacked and assume that the first (and only) entry is either package or
whatever we want. Then move the contents up a level and clean it up.